+++ CHANGELOG.md
@@ -9,7 +9,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 ### Added
 
+- Added UniqueValidator to check if a collection only contains unique values.
+
 ### Changed
+
+- Removed duplicate {{value}} placeholder from GreaterOrEqualValidator::NOT_GREATER_OR_EQUAL template.
 
 ### Deprecated